chore: rename Pascal Case folder of components
