+++ src/components/Avatar/AvatarPersonal.tsx
@@ -1,10 +1,6 @@
 import { Avatar } from '@mui/material';
 
-interface AvatarPersonalProps {
-  srcImage: string;
-}
-
-const AvatarPersonal = ({ srcImage }: AvatarPersonalProps): JSX.Element => {
+const AvatarPersonal = ({ srcImage }: { srcImage: string }): JSX.Element => {
   return <Avatar src={srcImage} alt='' sx={{ width: 300, height: 300 }} />;
 };
