+++ src/pages/SignUp/multi-step-form-state/reducer.ts
